split b ig energy bills controller into 2
