+++ app/javascript/controllers/energy_bill_upload_controller.js
@@ -1,17 +1,13 @@
 import { Controller } from "@hotwired/stimulus"
-import Dropzone from "dropzone"
 
 export default class extends Controller {
   static targets = [
     "title",
-    "btnContinue",
-    "btnDisplayFileDialog",
-    "btnAddMoreFiles",
-    "btnSubmit",
+    "lblHint",
     "dropZone",
-    "lblFilesAddedNow",
-    "filesAddedBefore",
-    "lblHint"
+    "btnContinue",
+    "btnAddMoreFiles",
+    "btnSubmit"
   ]
 
   static values = {
@@ -21,19 +17,12 @@ export default class extends Controller {
     pageTwoContinueButton: String,
     pageThreeTitle: String,
     pageThreeContinueButton: String,
-    addFileUrl: String,
-    removeFileUrl: String
   }
 
-  connect() {
-    this.setupDropZone()
-  }
+  static outlets = ["dropzone"]
 
-  initialize() {
-    this.dropzone = undefined
-    this.fileUploadErrors = {}
-    this.filesToUpload = []
-    this.uploadedFiles = []
+  dropzoneOutletConnected(outlet) {
+    outlet.setupOnUploadFilesComplete(this.continueToPageThree.bind(this))
   }
 
   // Actions
@@ -54,7 +43,7 @@ export default class extends Controller {
 
   uploadFilesOrContinueToPageThree() {
     if (this.anyFilesQueuedForUpload())
-      this.dropzone.processQueue() // moves the page 3 when done
+      this.uploadFiles() // move to page 3 when done
     else
       this.continueToPageThree()
   }
@@ -72,7 +61,7 @@ export default class extends Controller {
     this.btnAddMoreFilesTarget.blur()
 
     this.pageOne()
-    this.display(this.filesAddedBeforeTarget, true)
+    this.display(this.dropzoneOutlet.filesAddedBeforeTarget, true)
     this.moveFilesFromAddedToAlreadyUploaded()
   }
 
@@ -86,24 +75,24 @@ export default class extends Controller {
     this.titleTarget.innerHTML = this.pageOneTitleValue
     this.btnContinueTarget.innerHTML = this.pageOneContinueButtonValue
 
+    this.display(this.lblHintTarget, true)
+    this.display(this.dropZoneTarget, true)
+    this.display(this.dropzoneOutlet.lblFilesAddedNowTarget, false)
+    this.display(this.dropzoneOutlet.filesAddedBeforeTarget, false)
     this.display(this.btnContinueTarget, true)
     this.display(this.btnSubmitTarget, false)
-    this.display(this.dropZoneTarget, true)
-    this.display(this.lblHintTarget, true)
-    this.display(this.lblFilesAddedNowTarget, false)
-    this.display(this.filesAddedBeforeTarget, false)
     this.display(this.btnAddMoreFilesTarget, false)
   }
 
   pageTwo() {
     this.titleTarget.innerHTML = this.pageTwoTitleValue
 
+    this.display(this.lblHintTarget, false)
+    this.display(this.dropZoneTarget, false)
+    this.display(this.dropzoneOutlet.lblFilesAddedNowTarget, false)
+    this.display(this.dropzoneOutlet.filesAddedBeforeTarget, false)
     this.display(this.btnContinueTarget, false)
     this.display(this.btnSubmitTarget, false)
-    this.display(this.dropZoneTarget, false)
-    this.display(this.lblHintTarget, false)
-    this.display(this.lblFilesAddedNowTarget, false)
-    this.display(this.filesAddedBeforeTarget, false)
     this.display(this.btnAddMoreFilesTarget, false)
   }
 
@@ -111,12 +100,12 @@ export default class extends Controller {
     this.titleTarget.innerHTML = this.pageThreeTitleValue
     this.btnContinueTarget.innerHTML = this.pageThreeContinueButtonValue
 
+    this.display(this.lblHintTarget, false)
+    this.display(this.dropZoneTarget, false)
+    this.display(this.dropzoneOutlet.lblFilesAddedNowTarget, false)
+    this.display(this.dropzoneOutlet.filesAddedBeforeTarget, false)
     this.display(this.btnContinueTarget, false)
     this.display(this.btnSubmitTarget, true)
-    this.display(this.dropZoneTarget, false)
-    this.display(this.lblHintTarget, false)
-    this.display(this.lblFilesAddedNowTarget, false)
-    this.display(this.filesAddedBeforeTarget, false)
     this.display(this.btnAddMoreFilesTarget, true)
   }
 
@@ -127,167 +116,29 @@ export default class extends Controller {
       element.classList.add('govuk-!-display-none')
   }
 
-  // Dropzone
-
-  setupDropZone() {
-    const container = this.element.querySelector('.drop-zone-container')
-    const previewsContainer = container.querySelector('.file-previews .files-added-now')
-
-    const previewTemplateContainer = container.querySelector('.upload-preview-template')
-    const previewTemplate = previewTemplateContainer.innerHTML
-    previewTemplateContainer.remove()
-
-    this.dropzone = new Dropzone(
-      container,
-      {
-        url: this.addFileUrlValue,
-        parallelUploads: 3,
-        previewTemplate,
-        previewsContainer,
-        autoProcessQueue: false,
-        clickable: this.btnDisplayFileDialogTarget,
-        maxFilesize: 200
-      }
-    )
-
-    const controller = this
-
-    this.dropzone.on('addedfile', function(file) {
-      controller.onFileAdded(file)
-    })
-
-    this.dropzone.on('error', function(file, error) {
-      controller.onFileError(file, error)
-    })
-
-    this.dropzone.on('removedfile', function(file) {
-      controller.onFileRemoved(file)
-    })
-
-    this.dropzone.on('uploadprogress', function(file, progress) {
-      controller.onUploadProgress(file, progress)
-    })
-
-    this.dropzone.on('complete', function(file) {
-      controller.onFileUploadComplete(file)
-    })
-
-    this.dropzone.on('queuecomplete', function() {
-      controller.onQueueComplete()
-    })
+  uploadFiles() {
+    this.dropzoneOutlet.uploadFiles()
   }
-
-  // Dropzone events
-
-  onFileError(file, error) {
-    console.log("onFileError", file, error)
-    this.display(file.previewElement.querySelector('.progress-status-container'), true)
-    this.fileUploadErrors[file] = error
-  }
-
-  onFileAdded(file) {
-    this.display(this.lblFilesAddedNowTarget, true)
-    this.filesToUpload.push(file)
-
-    console.log("onFileAdded", this.filesToUpload, file)
-  }
-
-  onFileRemoved(file) {
-    if (this.fileHasBeenUploaded(file)) {
-      this.deleteFileFromServer(file)
-    } else {
-      this.deleteFileFromUi(file)
-    }
-  }
-
-  onUploadProgress(file, progress) {
-    const status = progress === 100 ? 'Complete' : 'Uploading'
-
-    const statusContainer = file.previewElement.querySelector('.progress-status')
-    statusContainer.classList.remove("uploading", "complete")
-    statusContainer.classList.add(status.toLowerCase())
-    statusContainer.innerHTML = status
-
-    file.previewElement.querySelector('[data-dz-uploadprogress]').style.width = progress
-
-    for (let hiddenElement of file.previewElement.querySelectorAll(".govuk-\\!-display-none")) {
-      this.display(hiddenElement, true)
-    }
-  }
-
-  onFileUploadComplete(file) {
-    console.log("onFileUploadComplete", file)
-
-    if (file.status === "success") {
-      this.filesToUpload = this.filesToUpload.filter(f => f != file)
-
-      const { energy_bill_id } = JSON.parse(file.xhr.response)
-      file.energy_bill_id = energy_bill_id
-      this.uploadedFiles.push(file)
-    } else {
-      // TODO: handle file error!!
-    }
-  }
-
-  onQueueComplete() {
-    this.continueToPageThree()
-  }
-
-  // Event action helpers
-
-  deleteFileFromServer(file) {
-    const body = new FormData()
-    body.append('id', file.energy_bill_id)
-
-    fetch(this.removeFileUrlValue, { method: 'DELETE', body })
-      .then(() => this.deleteFileFromUi(file))
-      .catch((error) => alert(error))
-  }
-
-  deleteFileFromUi(file) {
-    this.filesToUpload = this.filesToUpload.filter(pendingFile => pendingFile != file)
-    delete this.fileUploadErrors[file]
-
-    this.display(this.lblFilesAddedNowTarget, this.anyFilesQueuedForUpload())
-  }
-
-  // Dropzone presentational
 
   moveFilesFromAlreadyUploadedToAdded() {
-    const container = this.element.querySelector('.drop-zone-container')
-    const alreadyUploaded = container.querySelector('.files-added-before')
-    const filesAdded = container.querySelector('.files-added-now')
-
-    for (let uploadedFile of alreadyUploaded.querySelectorAll('.upload-item')) {
-      filesAdded.appendChild(uploadedFile)
-    }
+    this.dropzoneOutlet.moveFilesFromAlreadyUploadedToAdded()
   }
 
   moveFilesFromAddedToAlreadyUploaded() {
-    const container = this.element.querySelector('.drop-zone-container')
-    const alreadyUploaded = container.querySelector('.files-added-before')
-    const filesAdded = container.querySelector('.files-added-now')
-
-    for (let uploadedFile of filesAdded.querySelectorAll('.upload-item')) {
-      alreadyUploaded.appendChild(uploadedFile)
-    }
+    this.dropzoneOutlet.moveFilesFromAddedToAlreadyUploaded()
   }
 
   // Query methods
 
   anyFilesQueuedForUpload() {
-    return this.filesToUpload.length > 0
+    return this.dropzoneOutlet.anyFilesQueuedForUpload()
   }
 
   anyFilesUploadedSuccessfully() {
-    return this.uploadedFiles.length > 0
+    return this.dropzoneOutlet.anyFilesUploadedSuccessfully()
   }
 
   anyFileUploadErrorsOccured() {
-    return Object.keys(this.fileUploadErrors).length > 0
-  }
-
-  fileHasBeenUploaded(file) {
-    return file.energy_bill_id !== undefined
+    return this.dropzoneOutlet.anyFileUploadErrorsOccured()
   }
 }
